Split checker.rs into focused sub-modules
Reduces checker.rs from 2453 to 627 lines by extracting method
groups into 6 new sub-modules:

- imports.rs (191 lines): check_import, lookup_resolved_symbol
- type_registration.rs (451 lines): register_type_decl, flatten_record_spreads
- type_compat.rs (244 lines): types_compatible, types_unifiable, merge_types
- traits.rs (85 lines): register_trait_decl, check_trait_decl, check_trait_impl
- type_resolve.rs (176 lines): resolve_type, resolve_named_type
- items.rs (706 lines): check_const, check_function, check_for_block, JSX, probes

Purely structural - no behavioral changes.
